oprava #31 - tvari se to, ze progressbary jsou uz ok.

+++ trunk/AndariaPatcher/src/cz/polous/andaria/Downloader.java
@@ -51,7 +51,7 @@ class Downloader extends PatcherQueue {
         PatchItem p = getFirstItem();
         // ProgressBars
 
-        resetSingleProgress((double) p.getSize());
+        resetSingleProgress(p.getSize());
 
         // Check if wanted file exists at local storage
         String fileName = p.getLocalFileName();
